move to __init__.py, correct async usage
